requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.19.0
google-generativeai>=0.8.0
google-genai
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
requests>=2.32.0
httpx>=0.27.0
gspread>=6.1.0
oauth2client>=4.1.3
numpy>=1.26.0
tenacity
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
google-auth
